feat: Add the finalised forestation VIEW
Something that wasn't covered during the SQL course on Udacity was how
to actually use a VIEW once it has been created. First off, a user needs
to run the entire query including the CREATE VIEW portion. Once this is
complete, then a user may use the virtual table created by the VIEW as
if it was an existing table in the database.