fix: Modify forestation VIEW and test Question 1(d)
I made a mistake in my initial forestation VIEW table. I rounded off the
forest area percentage to the nearest unit. This wasn't a problem in
general, but it caused an issue to be raised by the Project Reviewer
when looking at Question 3(c), as the code was not correctly assigning
the countries into their correct percentiles because the values involved
were not rounded off to decimal places.

Also, I need to test all the other queries again as I modified the view.
Question 1(d), the SQL query had to be modified to produce a better
looking result.